backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
    "psycopg2-binary>=2.9.9",
    "fastapi[standard]>=0.128.0",
    "playwright>=1.57.0",
    "requests>=2.32.0",
    "google-genai>=1.58.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "ruff>=0.14.11",
]

[tool.ruff.lint]
# On top of the default `select` (`E4`, E7`, `E9`, and `F`), enable sorting imports.
extend-select = ["I", "UP"]
ignore = ["UP028"]
